third_party/blink/web_tests/http/tests/security/referrer-policy-redirect-link.html

<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFrames();
    testRunner.waitUntilDone();
    testRunner.setPopupBlockingEnabled(false);
}

function runTest() {
    var iframe = document.getElementById("iframe");
    iframe.contentWindow.postMessage({"action": "click"}, "*");
}
</script>
</head>
<body>
<p>
This test checks the referrer policy is obeyed along the redirect chain.  The
test passes if the referrer is https://127.0.0.1:8443
</p>
<iframe id="iframe" name="iframe" onload="runTest()" src="https://127.0.0.1:8443/security/resources/referrer-policy-redirect-link.html"></iframe>
</body>
</html>
